requirements.txt
asgiref==3.3.1
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
Faker==5.0.1
Markdown==3.3.3
mysql-connector-python==8.0.22
mysqlclient==2.0.2
Pillow==8.0.1
pkg-resources==0.0.0
protobuf==3.14.0    
PyMySQL==0.10.1
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.15.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
  